Don't process unwanted attributes. Corrected dict history naming

+++ fetch-ccm-history.py
@@ -13,7 +13,7 @@ import SynergySession
 import FileObject
 import TaskObject
 import SynergyObject
-from SynergyUtils import ObjectHistory, TaskUtil
+from SynergyUtils import ObjectHistory, TaskUtil, CCMFilePath, SynergyUtils
 
 from operator import itemgetter, attrgetter
 
@@ -25,7 +25,7 @@ import sys
 
 class CCMHistory(object):
     """Get History (objects and tasks) in a Synergy (ccm) database between baseline projects"""
-    
+
     def __init__(self, ccm, history, outputfile):
         self.ccm = ccm
         self.delim = self.ccm.delim()
@@ -46,53 +46,76 @@ class CCMHistory(object):
                 latestproject = SynergyObject.SynergyObject(s['objectname'], self.delim, s['owner'], s['status'], s['create_time'], s['task'])
                 self.tag = s['version']
         print "Latest project:", latestproject.get_object_name(), "created:", latest
-        
+
         #find baseline of latestproject:
         base = self.ccm.query("is_baseline_project_of('{0}')".format(latestproject.get_object_name())).format("%objectname").format("%create_time").format('%version').format("%owner").format("%status").format("%task").run()[0]
         baseline_project = SynergyObject.SynergyObject(base['objectname'], self.delim, base['owner'], base['status'], base['create_time'], base['task'])
         print "Baseline project:", baseline_project.get_object_name()
-        
+        if self.tag not in self.history.keys():
+            self.history[self.tag] = {'objects': [], 'tasks': []}
+        self.history[self.tag]['next'] = None
+
         while baseline_project:
             print "Toplevel Project:", latestproject.get_object_name()
             # do the history thing
             self.create_history(latestproject.get_object_name(), baseline_project.get_object_name())
             self.history[self.tag]['created'] = latestproject.get_created_time()
-            
+
+            next = latestproject.get_version()
+
             # Find next baseline project
             latestproject = baseline_project
             baseline = self.ccm.query("is_baseline_project_of('{0}')".format(latestproject.get_object_name())).format("%objectname").format("%create_time").format('%version').format("%owner").format("%status").format("%task").run()[0]
             baseline_project = SynergyObject.SynergyObject(baseline['objectname'], self.delim, baseline['owner'], baseline['status'], baseline['create_time'], baseline['task'])
 
             #Set previous project and name of current release:
-            self.history[self.tag]['previous'] = baseline_project.get_version()
+            self.history[self.tag]['previous'] = latestproject.get_version()
             self.history[self.tag]['name'] = self.tag
 
             #Store data
             fname = self.outputfile + '_' + self.tag
             self.persist_data(fname, self.history[self.tag])
-            # delete the _inc file
-            os.remove(fname + '_inc' + '.p')
+            # delete the _inc file if it exists
+            if os.path.isfile(fname + '_inc' + '.p'):
+                os.remove(fname + '_inc' + '.p')
+
+            #Print Info
+            print self.tag, "done processing, Info:"
+            print "Name        ", self.tag
+            print "Previous <- ", self.history[self.tag]['previous']
+            print "Next   ->   ", self.history[self.tag]['next']
+            print "Number of:  "
+            print "    Tasks:  ", str(len(self.history[self.tag]['tasks']))
+            print "    Files:  ", str(len(self.history[self.tag]['objects']))
+            print ""
 
             #Finally set the new (old) tag for the release
-            self.tag = baseline_project.get_version()
+            self.tag = latestproject.get_version()
+            if self.tag not in self.history.keys():
+                self.history[self.tag] = {'objects': [], 'tasks': []}
+            self.history[self.tag]['next'] = next
+
+
+
 
             # Stop at this release and get all objects from Synergy
-            if latestproject.get_version() == '10w45_sb9_fam':
-                self.history[self.tag] = self.get_project_objects(latestproject.get_object_name())
+            if latestproject.get_version() == '10w51_sb9_fam':
+                
                 self.history[self.tag]['previous'] = None
                 self.history[self.tag]['name'] = self.tag
                 self.history[self.tag]['created'] = latestproject.get_created_time()
+                self.history[self.tag] = self.get_project_objects(latestproject.get_object_name())
 
                 baseline_project = None
 
         return self.history
-            
-                
+
+
     def create_history(self, latestproject, baseline_project):
         #clear changed objects and find all objects from this release
         self.find_project_diff(latestproject, baseline_project, latestproject)
 
-    
+
     def find_project_diff(self, latestproject, baseline_project, toplevel_project):
         #print "diffenence between", latestproject, "and", baseline_project
         object_hist = ObjectHistory(self.ccm, toplevel_project)
@@ -107,7 +130,7 @@ class CCMHistory(object):
             self.history[self.tag] = {'objects': [], 'tasks': []}
         for o in objects_changed:
             #print o['objectname']
-            if o['objectname'] not in existing_objects:                
+            if o['objectname'] not in existing_objects:
                 if ':project:' in o['objectname']:
                     #print o['objectname']
                     #Find diffenence of subprojects
@@ -130,15 +153,16 @@ class CCMHistory(object):
                         else:
                             print "no predecessor..."
                     #print ""
-                        
+
                 else:
                     # Get history for this file between these to baselines/ projects
                     objects.update(object_hist.get_history(FileObject.FileObject(o['objectname'], self.delim, o['owner'], o['status'], o['create_time'], o['task']), baseline_project))
                     persist = True
                     #self.changed_objects.update(object_hist.get_history(FileObject.FileObject(o['objectname'], self.delim, o['owner'], o['status'], o['create_time'], o['task']), baseline_project))
-        
+            else:
+                print o['objectname'], "already in history"
         # Create tasks from objects
-        self.find_tasks_from_objects(objects.values(), latestproject)        
+        self.find_tasks_from_objects(objects.values(), latestproject)
         #persist data
         self.history[self.tag]['objects'].extend(objects.values())
         if persist:
@@ -149,18 +173,18 @@ class CCMHistory(object):
     def find_tasks_from_objects(self, objects, project):
         task_util = TaskUtil(self.ccm)
         tasks = {}
-        not_used = []            
+        not_used = []
         #Find all tasks from the objects found
         for o in objects:
             for task in o.get_tasks().split(','):
                 if task != "<void>":
                     if task not in tasks.keys():
                         if task not in not_used:
-                            # create task object 
+                            # create task object
                             print "Task:", task
                             if task_util.task_in_project(task, project):
                                 result = self.ccm.query("name='{0}' and instance='{1}'".format('task' + task.split('#')[1], task.split('#')[0])).format("%objectname").format("%owner").format("%status").format("%create_time").format("%task_synopsis").format("%release").run()
-                            
+
                                 t = result[0]
                                 # Only use completed tasks!
                                 if t['status'] == 'completed':
@@ -168,13 +192,13 @@ class CCMHistory(object):
                                     to.set_synopsis(t['task_synopsis'])
                                     to.set_release(t['release'])
                                     to.add_object(o)
-                                    
+
                                     tasks[task] = to
                             else:
                                 not_used.append(task)
                     else:
                         tasks[task].add_object(o)
-                            
+
         # Fill out all task info
         for task in tasks.values():
             task_util.fill_task_info(task)
@@ -183,7 +207,8 @@ class CCMHistory(object):
 
 
     def get_project_objects(self, project):
-        objects_changed = self.ccm.query("recursive_is_member_of('{0}')".format(project)).format("%objectname").format("%owner").format("%status").format("%create_time").format("%task").run()
+        object_hist = ObjectHistory(self.ccm, project)
+        objects_changed = self.ccm.query("is_member_of('{0}')".format(project)).format("%objectname").format("%owner").format("%status").format("%create_time").format("%task").run()
         objects = {}
         existing_objects = []
         persist = False
@@ -192,15 +217,24 @@ class CCMHistory(object):
                 existing_objects = [o.get_object_name() for o in self.history[self.tag]['objects']]
         else:
             self.history[self.tag] = {'objects': [], 'tasks': []}
+
         for o in objects_changed:
             #print o['objectname']
             if o['objectname'] not in existing_objects:
-                if ':project:' not in o['objectname']:
-                    # Get information for this file  - no history
+                if ':project:' in o['objectname']:
+                    #print o['objectname']
+                    #Find diffenence of subprojects
+                    #first find subprojects baseline project
+                    subproject = o['objectname']
+                    print "Subproject", subproject
+                    self.get_project_objects(subproject)
+
+                else:
                     objects.update(object_hist.get_history(FileObject.FileObject(o['objectname'], self.delim, o['owner'], o['status'], o['create_time'], o['task']), project))
                     persist = True
                     #self.changed_objects.update(object_hist.get_history(FileObject.FileObject(o['objectname'], self.delim, o['owner'], o['status'], o['create_time'], o['task']), baseline_project))
-
+            else:
+                print o['objectname'], "already in history"
         # Create tasks from objects
         self.find_tasks_from_objects(objects.values(), project)
         #persist data
@@ -217,40 +251,42 @@ class CCMHistory(object):
         cPickle.dump(data, fh)
         fh.close()
         print "done..."
-    
-               
+
+
 
 def main():
     # make all stdout stderr writes unbuffered/ instant/ inline
     sys.stdout =  os.fdopen(sys.stdout.fileno(), 'w', 0);
     sys.stderr =  os.fdopen(sys.stderr.fileno(), 'w', 0);
-    
+
     ccm_db = sys.argv[1]
     project = sys.argv[2]
     outputfile = sys.argv[3]
-    
+
     print "Starting Synergy session on", ccm_db, "..."
     ccm = SynergySession.SynergySession(ccm_db)
     print "session started"
     delim = ccm.delim()
     history = {}
-    fname = outputfile
+    fname = outputfile + '.p'
     if os.path.isfile(fname):
+        print "Loading", fname, "..."
         fh = open(fname, 'rb')
         history = cPickle.load(fh)
         fh.close()
-        
+        print "history contains:", history.keys()
+
     #print history
     fetch_ccm = CCMHistory(ccm, history, outputfile)
-    
+
     history = fetch_ccm.get_project_history(project)
-    
-    
+
+
     fh = open(outputfile + '.p', 'wb')
     cPickle.dump(history, fh)
     fh.close()
-    
-    
-    
+
+
+
 if __name__ == '__main__':
     main()
